+++ packages/terra-menu/src/Menu.jsx
@@ -53,12 +53,17 @@ const propTypes = {
    * Otherwise, the menu will display without an arrow and right aligned.
    */
   isArrowDisplayed: PropTypes.bool,
+  /**
+   * Header Title for menu if menu contains drill-in sub-menus
+   */
+  headerTitle: PropTypes.string,
 };
 
 const defaultProps = {
   isArrowDisplayed: false,
   isOpen: false,
   contentWidth: '240',
+  headerTitle: '',
 };
 
 class Menu extends React.Component {
@@ -144,6 +149,7 @@ class Menu extends React.Component {
         index={index}
         boundingRef={boundingRef}
         isFocused={index === visiblePage}
+        headerTitle={this.props.headerTitle}
       >
         {item.props.children || item.props.subMenuItems}
       </MenuContent>